tambah seeder + debugbar

+++ Modules/User/Http/Controllers/UsersController.php
@@ -17,7 +17,7 @@ use Yajra\DataTables\DataTables;
 class UsersController extends Controller
 {
     public function index() {
-        // abort_if(Gate::denies('access_user_management'), 403);
+        abort_if(Gate::denies('access_user_management'), 403);
 
         $users = User::all();
 
@@ -27,7 +27,7 @@ class UsersController extends Controller
 
     public function create()
     {
-        // abort_if(Gate::denies('access_user_management'), 403);
+        abort_if(Gate::denies('access_user_management'), 403);
 
         return view('user::users.create');
     }
@@ -35,7 +35,7 @@ class UsersController extends Controller
 
     public function store(Request $request)
     {
-        // abort_if(Gate::denies('access_user_management'), 403);
+        abort_if(Gate::denies('access_user_management'), 403);
 
         $request->validate([
             'name'     => 'required|string|max:255',
@@ -63,7 +63,7 @@ class UsersController extends Controller
             }
         }
 
-        toast("User Created & Assigned '$request->role' Role!", 'success');
+        // toast("User Created & Assigned '$request->role' Role!", 'success');
 
         return redirect()->route('users.index');
     }
@@ -71,7 +71,7 @@ class UsersController extends Controller
 
     public function edit(User $user)
     {
-        // abort_if(Gate::denies('access_user_management'), 403);
+        abort_if(Gate::denies('access_user_management'), 403);
 
         return view('user::users.edit', compact('user'));
     }
@@ -79,7 +79,7 @@ class UsersController extends Controller
 
     public function update(Request $request, User $user)
     {
-        // abort_if(Gate::denies('access_user_management'), 403);
+        abort_if(Gate::denies('access_user_management'), 403);
 
         $request->validate([
             'name'     => 'required|string|max:255',
@@ -109,7 +109,7 @@ class UsersController extends Controller
             }
         }
 
-        toast("User Updated & Assigned '$request->role' Role!", 'info');
+        // toast("User Updated & Assigned '$request->role' Role!", 'info');
 
         return redirect()->route('users.index');
     }
@@ -117,11 +117,11 @@ class UsersController extends Controller
 
     public function destroy(User $user)
     {
-        // abort_if(Gate::denies('access_user_management'), 403);
+        abort_if(Gate::denies('access_user_management'), 403);
 
         $user->delete();
 
-        toast('User Deleted!', 'warning');
+        // toast('User Deleted!', 'warning');
 
         return redirect()->route('users.index');
     }
